moved Interface.ts from Utils to types folder
